chore: rename entities to models
