+++ books-msvc/src/models/book.model.ts
