+++ books-msvc/src/resolvers/book.resolver.ts
@@ -2,7 +2,7 @@ import { Arg, Mutation, Query, Resolver } from 'type-graphql';
 import { ObjectId } from 'mongodb';
 import { Promise } from 'mongoose';
 
-import { Book, BookModel } from '../entities/book.model';
+import { Book, BookModel } from '../models/book.model';
 import { BookInput } from '../types/book.input';
 import { ObjectIdScalar } from '../scalars/object-id.scalar';
